# apps/gateway/src/swagger/swagger.doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contact/import-google-contacts": {
      "post": {
        "operationId": "ContactController_importContactsFromGoogle",
        "summary": "Import contacts from google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportContactsFromGoogleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/get-user-contacts-with-filter": {
      "post": {
        "operationId": "ContactController_getUserContacts",
        "summary": "Get all contacts of user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserContactsWithFilterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserContactsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/add-to-favourite": {
      "post": {
        "operationId": "ContactController_addContactToFavourite",
        "summary": "Add contact to favourite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactToFavouriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/remove-from-favourite": {
      "post": {
        "operationId": "ContactController_removeContactFromFavourite",
        "summary": "Remove contact from favourite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactToFavouriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/create-contact": {
      "post": {
        "operationId": "ContactController_createNewContact",
        "summary": "Create new contact",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewContactRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/change-status-of-contact": {
      "post": {
        "operationId": "ContactController_changeStatusOfContact",
        "summary": "Change status of contact",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusOfContactRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/create-custom-group": {
      "post": {
        "operationId": "ContactController_createCustomGroup",
        "summary": "Delete contact",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactCustomGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContactCustomGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/add-contact-to-group": {
      "post": {
        "operationId": "ContactController_addContactToGroup",
        "summary": "Add contact to group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactToGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/delete-member-from-group": {
      "post": {
        "operationId": "ContactController_deleteMemberFromGroup",
        "summary": "Delete member from group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMemberFromGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/add-to-document-group": {
      "post": {
        "operationId": "ContactController_addContactToDocumentGroup",
        "summary": "Add contact to group of document",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactToDocumentGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/delete-from-document-group": {
      "post": {
        "operationId": "ContactController_deleteContactFromDocumentGroup",
        "summary": "Delete from document group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteContactFromDocumentGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/delete-contact-group": {
      "post": {
        "operationId": "ContactController_deleteContactGroup",
        "summary": "Delete contact group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactGroupIdRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/update-contact-group": {
      "post": {
        "operationId": "ContactController_updateContactGroup",
        "summary": "Update contact group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/import-from-csv": {
      "post": {
        "operationId": "ContactController_importContactsFromCsv",
        "summary": "Import contacts from csv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImportContactsFromCsvRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/csv-example": {
      "get": {
        "operationId": "ContactController_getCsvExample",
        "summary": "Get CSV example file url",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCsvExampleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/update-contact": {
      "post": {
        "operationId": "ContactController_updateContact",
        "summary": "Update contact",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/groups-and-contacts": {
      "post": {
        "operationId": "ContactController_getAllDocumentGroupsAndContacts",
        "summary": "Get document groups and contacts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllDocumentGroupsAndContactsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDocumentGrroupsAndContactsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/groups-with-contacts": {
      "get": {
        "operationId": "ContactController_getGroupsWithContacts",
        "summary": "Get groups with contacts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDocumentGrroupsAndContactsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/add-group-to-favourite": {
      "post": {
        "operationId": "ContactController_addGroupToFavourite",
        "summary": "Add group to favourites",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupToFavouriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/contact/change-status-of-contacts-bulk": {
      "post": {
        "operationId": "ContactController_changeStatusOfContactsBulk",
        "summary": "Change status of contacts bulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusOfContactsBulkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "",
    "description": "\"Rentil\" - API documentation.<br><br>Link to <a  target='_blank' href=''>API Guide</a>",
    "version": "1.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "{schema}://undefined",
      "description": "API Server",
      "variables": {
        "schema": {
          "enum": [
            "https"
          ],
          "default": "https"
        }
      }
    },
    {
      "url": "{schema}://localhost:3000",
      "description": "Local API Server",
      "variables": {
        "schema": {
          "enum": [
            "http"
          ],
          "default": "http"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X_API_KEY": {
        "type": "apiKey",
        "name": "X-API-Key",
        "description": "Application key",
        "in": "header"
      },
      "X_SESSION_KEY": {
        "type": "apiKey",
        "name": "X-Session-Key",
        "description": "Application key",
        "in": "header"
      }
    },
    "schemas": {
...
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "ImportContactsFromGoogleRequest": {
        "type": "object",
        "properties": {
          "userToken": {
            "type": "string"
          }
        },
        "required": [
          "userToken"
        ]
      },
...
      "GetUserContactsResponse": {
        "type": "object",
        "properties": {
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignyBaseInfo"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetGroupsWithSearchResponse"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfoResponse"
          }
        },
        "required": [
          "pageInfo"
        ]
      },
...
    }
  }
}